Fix entity AABB rotations
Order of rotations is important:
- When ray tracing we rotate the ray into the entity's frame of
  reference
- When preparing AABB's we rotate the AABB in the global frame of
  reference

Reversing the order of rotations for AABBs seems to resolve the bounding
box rotation not matching with the ray rotation, at least on my simple
tests. This matches the behavior of the old "transform wrapper"
approach, where transforms would be evaluated in reverse order as the
ray progressed through the chain of wrappers.